Add Gate for authorization

+++ app/Http/Controllers/Question/QuestionVoteController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Question;
 use App\Http\Controllers\Controller;
 use App\Models\QuestionVote;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionVoteController extends Controller
 {
@@ -16,13 +17,15 @@ class QuestionVoteController extends Controller
      */
     public function __invoke(Request $request)
     {
+        Gate::authorize('users-vote');
+
         $vote = QuestionVote::updateOrCreate(
             ['user_id' => auth()->id(), 'question_id' => $request->question_id],
             ['vote' => $request->vote]
         );
 
         if ($vote->wasRecentlyCreated === false) {
-            if (! $vote->wasChanged('vote')) {
+            if (!$vote->wasChanged('vote')) {
                 $vote->delete();
             }
         }
